ACA-68 - moved NoticeOfChangeService class back out of noc subdirectory
Moving this class now will mean continuing merge conflicts each time
changes in ACA-72 (and therefore ACA-71) are pulled in.  Moving back for
the time being
